+++ clients/admin-ui/src/features/common/Header.tsx
@@ -24,27 +24,27 @@ interface HeaderProps {
 const Header: React.FC<HeaderProps> = ({ username }) => (
   <header>
     <Flex
-      bg="gray.50"
-      width="100%"
+      bg='gray.50'
+      width='100%'
       py={3}
       px={10}
-      justifyContent="space-between"
-      alignItems="center"
+      justifyContent='space-between'
+      alignItems='center'
     >
-      <NextLink href="/" passHref>
+      <NextLink href='/' passHref>
         {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-        <Link display="flex">
-          <Image src="/logo.svg" width={83} height={26} alt="FidesOps Logo" />
+        <Link display='flex'>
+          <Image src='/logo.svg' width={83} height={26} alt='FidesOps Logo' />
         </Link>
       </NextLink>
-      <Flex alignItems="center">
+      <Flex alignItems='center'>
         <Menu>
-          <MenuButton as={Button} size="sm" variant="ghost">
-            <UserIcon color="gray.700" />
+          <MenuButton as={Button} size='sm' variant='ghost'>
+            <UserIcon color='gray.700' />
           </MenuButton>
-          <MenuList shadow="xl">
+          <MenuList shadow='xl'>
             <Stack px={3} py={2} spacing={0}>
-              <Text fontWeight="medium">{username}</Text>
+              <Text fontWeight='medium'>{username}</Text>
               {/* This text should only show if actually an admin */}
               {/* <Text fontSize="sm" color="gray.600">
                 Administrator