+++ clients/admin-ui/src/features/common/RequestStatusBadge.tsx
@@ -1,7 +1,7 @@
 import { Badge } from '@fidesui/react';
 import { ComponentProps } from 'react';
 
-import { PrivacyRequestStatus } from './types';
+import { PrivacyRequestStatus } from '../privacy-requests/types';
 
 export const statusPropMap: {
   [key in PrivacyRequestStatus]: ComponentProps<typeof Badge>;
@@ -40,16 +40,16 @@ interface RequestBadgeProps {
   status: keyof typeof statusPropMap;
 }
 
-const RequestBadge: React.FC<RequestBadgeProps> = ({ status }) => (
+const RequestStatusBadge: React.FC<RequestBadgeProps> = ({ status }) => (
   <Badge
-    color="white"
+    color='white'
     bg={statusPropMap[status].bg}
     width={107}
-    lineHeight="18px"
-    textAlign="center"
+    lineHeight='18px'
+    textAlign='center'
   >
     {statusPropMap[status].label}
   </Badge>
 );
 
-export default RequestBadge;
+export default RequestStatusBadge;