+++ clients/admin-ui/src/features/subject-request/ActivityTimeline.tsx
@@ -1,0 +1,72 @@
+import { Box, Flex, Text } from '@fidesui/react';
+import React from 'react';
+
+import { GreenCheckCircle } from '../common/Icon';
+import { PrivacyRequest } from '../privacy-requests/types';
+
+type ActivityTimelineProps = {
+  subjectRequest: PrivacyRequest;
+  setEventDetails: Function;
+};
+
+const ActivityTimeline = ({
+  subjectRequest,
+  setEventDetails,
+}: ActivityTimelineProps) => {
+  const { results } = subjectRequest;
+
+  const resultKeys = [
+    'postgres_example',
+    'postgres_example',
+    'postgres_example',
+  ];
+
+  const timelineEntries = resultKeys.map((key, index) => (
+    <Box key={key}>
+      <Flex alignItems='center' height={23} position='relative'>
+        <Box zIndex={1}>
+          <GreenCheckCircle />
+        </Box>
+        {index === resultKeys.length - 1 ? null : (
+          <Box
+            width='2px'
+            height='63px'
+            backgroundColor='gray.700'
+            position='absolute'
+            top='16px'
+            left='6px'
+            zIndex={0}
+          />
+        )}
+
+        <Text color='gray.600' fontWeight='500' fontSize='sm' ml={2}>
+          {key}
+        </Text>
+      </Flex>
+      <Text
+        cursor='pointer'
+        color='complimentary.500'
+        fontWeight='500'
+        fontSize='sm'
+        ml={6}
+        mb={7}
+        onClick={() => {
+          setEventDetails(results![key]);
+        }}
+      >
+        View Details
+      </Text>
+    </Box>
+  ));
+
+  return (
+    <Box width='100%'>
+      <Text color='gray.900' fontSize='md' fontWeight='500' mb={1}>
+        Activity Timeline
+      </Text>
+      {timelineEntries}
+    </Box>
+  );
+};
+
+export default ActivityTimeline;