+++ clients/admin-ui/src/features/user-management/NewUserForm.tsx
@@ -11,19 +11,19 @@ import {
   Stack,
   Text,
   useToast,
-} from "@fidesui/react";
-import { useFormik } from "formik";
-import type { NextPage } from "next";
-import NextLink from "next/link";
-import { useRouter } from "next/router";
-import React from "react";
+} from '@fidesui/react';
+import { useFormik } from 'formik';
+import type { NextPage } from 'next';
+import NextLink from 'next/link';
+import { useRouter } from 'next/router';
+import React from 'react';
 
-import { isErrorWithDetail, isErrorWithDetailArray } from "../common/helpers";
-import { userPrivilegesArray } from "../user/types";
+import { isErrorWithDetail, isErrorWithDetailArray } from '../common/helpers';
+import { userPrivilegesArray } from '../user/types';
 import {
   useCreateUserMutation,
   useUpdateUserPermissionsMutation,
-} from "../user/user.slice";
+} from '../user/user.slice';
 
 const useUserForm = () => {
   const [createUser] = useCreateUserMutation();
@@ -33,10 +33,10 @@ const useUserForm = () => {
 
   const formik = useFormik({
     initialValues: {
-      username: "",
-      first_name: "",
-      last_name: "",
-      password: "",
+      username: '',
+      first_name: '',
+      last_name: '',
+      password: '',
       scopes: [],
     },
     onSubmit: async (values) => {
@@ -49,8 +49,8 @@ const useUserForm = () => {
 
       const createUserResult = await createUser(userBody);
 
-      if ("error" in createUserResult) {
-        let errorMsg = "An unexpected error occurred. Please try again.";
+      if ('error' in createUserResult) {
+        let errorMsg = 'An unexpected error occurred. Please try again.';
         if (isErrorWithDetail(createUserResult.error)) {
           errorMsg = createUserResult.error.data.detail;
         } else if (isErrorWithDetailArray(createUserResult.error)) {
@@ -58,7 +58,7 @@ const useUserForm = () => {
           errorMsg = error.data.detail[0].msg;
         }
         toast({
-          status: "error",
+          status: 'error',
           description: errorMsg,
         });
         return;
@@ -68,15 +68,15 @@ const useUserForm = () => {
 
       const userWithPrivileges = {
         id: data ? data.id : null,
-        scopes: [...values.scopes, "privacy-request:read"],
+        scopes: [...values.scopes, 'privacy-request:read'],
       };
 
       const updateUserPermissionsResult = await updateUserPermissions(
         userWithPrivileges as { id: string }
       );
 
-      if (!("error" in updateUserPermissionsResult)) {
-        router.push("/user-management");
+      if (!('error' in updateUserPermissionsResult)) {
+        router.push('/user-management');
       }
     },
     validate: (values) => {
@@ -88,31 +88,31 @@ const useUserForm = () => {
       } = {};
 
       if (!values.username) {
-        errors.username = "Username is required";
+        errors.username = 'Username is required';
       }
 
       if (!values.password) {
-        errors.password = "Password is required";
+        errors.password = 'Password is required';
       }
 
       if (values.password.length < 8) {
-        errors.password = "Password must have at least eight characters.";
+        errors.password = 'Password must have at least eight characters.';
       }
 
       if (!/[0-9]/.test(values.password)) {
-        errors.password = "Password must have at least one number.";
+        errors.password = 'Password must have at least one number.';
       }
 
       if (!/[A-Z]/.test(values.password)) {
-        errors.password = "Password must have at least one capital letter.";
+        errors.password = 'Password must have at least one capital letter.';
       }
 
       if (!/[a-z]/.test(values.password)) {
-        errors.password = "Password must have at least one lowercase letter.";
+        errors.password = 'Password must have at least one lowercase letter.';
       }
 
       if (!/[\W]/.test(values.password)) {
-        errors.password = "Password must have at least one symbol.";
+        errors.password = 'Password must have at least one symbol.';
       }
 
       return errors;
@@ -131,29 +131,29 @@ const UserForm: NextPage = () => {
   return (
     <div>
       <main>
-        <Heading mb={4} fontSize="xl" colorScheme="primary">
+        <Heading mb={4} fontSize='xl' colorScheme='primary'>
           Profile
         </Heading>
         <Divider mb={7} />
         <chakra.form
           onSubmit={handleSubmit}
-          maxW={["xs", "xs", "100%"]}
-          width="100%"
+          maxW={['xs', 'xs', '100%']}
+          width='100%'
         >
           <Stack mb={8} spacing={6}>
             <FormControl
-              id="username"
+              id='username'
               isInvalid={touched.username && Boolean(errors.username)}
             >
-              <FormLabel htmlFor="username" fontWeight="medium">
+              <FormLabel htmlFor='username' fontWeight='medium'>
                 Username
               </FormLabel>
               <Input
-                id="username"
-                maxWidth="40%"
-                name="username"
-                focusBorderColor="primary.500"
-                placeholder="Enter new username"
+                id='username'
+                maxWidth='40%'
+                name='username'
+                focusBorderColor='primary.500'
+                placeholder='Enter new username'
                 onChange={handleChange}
                 onBlur={handleBlur}
                 value={values.username}
@@ -162,51 +162,51 @@ const UserForm: NextPage = () => {
               <FormErrorMessage>{errors.username}</FormErrorMessage>
             </FormControl>
 
-            <FormControl id="first_name">
-              <FormLabel htmlFor="first_name" fontWeight="medium">
+            <FormControl id='first_name'>
+              <FormLabel htmlFor='first_name' fontWeight='medium'>
                 First Name
               </FormLabel>
               <Input
-                id="first_name"
-                maxWidth="40%"
-                name="first_name"
-                focusBorderColor="primary.500"
-                placeholder="Enter first name of user"
+                id='first_name'
+                maxWidth='40%'
+                name='first_name'
+                focusBorderColor='primary.500'
+                placeholder='Enter first name of user'
                 onChange={handleChange}
                 onBlur={handleBlur}
                 value={values.first_name}
               />
             </FormControl>
 
-            <FormControl id="last_name">
-              <FormLabel htmlFor="last_name" fontWeight="medium">
+            <FormControl id='last_name'>
+              <FormLabel htmlFor='last_name' fontWeight='medium'>
                 Last Name
               </FormLabel>
               <Input
-                id="last_name"
-                maxWidth="40%"
-                name="last_name"
-                focusBorderColor="primary.500"
-                placeholder="Enter last name of user"
+                id='last_name'
+                maxWidth='40%'
+                name='last_name'
+                focusBorderColor='primary.500'
+                placeholder='Enter last name of user'
                 onChange={handleChange}
                 onBlur={handleBlur}
                 value={values.last_name}
               />
             </FormControl>
             <FormControl
-              id="password"
+              id='password'
               isInvalid={touched.password && Boolean(errors.password)}
             >
-              <FormLabel htmlFor="password" fontWeight="medium">
+              <FormLabel htmlFor='password' fontWeight='medium'>
                 Password
               </FormLabel>
               <Input
-                id="password"
-                maxWidth="40%"
-                name="password"
-                focusBorderColor="primary.500"
-                placeholder="********"
-                type="password"
+                id='password'
+                maxWidth='40%'
+                name='password'
+                focusBorderColor='primary.500'
+                placeholder='********'
+                type='password'
                 value={values.password}
                 onChange={handleChange}
                 onBlur={handleBlur}
@@ -215,30 +215,30 @@ const UserForm: NextPage = () => {
               <FormErrorMessage>{errors.password}</FormErrorMessage>
             </FormControl>
             <Divider mb={7} mt={7} />
-            <Heading fontSize="xl" colorScheme="primary">
+            <Heading fontSize='xl' colorScheme='primary'>
               Privileges
             </Heading>
             <Text>Select privileges to assign to this user</Text>
             <Divider mb={2} mt={2} />
 
-            <Stack spacing={[1, 5]} direction="column">
+            <Stack spacing={[1, 5]} direction='column'>
               {userPrivilegesArray.map((policy) => (
                 <Checkbox
-                  colorScheme="purple"
-                  defaultChecked={policy.scope === "privacy-request:read"}
+                  colorScheme='purple'
+                  defaultChecked={policy.scope === 'privacy-request:read'}
                   key={`${policy.privilege}`}
                   onChange={handleChange}
                   id={`scopes-${policy.privilege}`}
-                  name="scopes"
+                  name='scopes'
                   // @ts-ignore
                   isChecked={values.scopes[policy.privilege]}
                   value={
-                    policy.scope === "privacy-request:read"
+                    policy.scope === 'privacy-request:read'
                       ? undefined
                       : policy.scope
                   }
-                  isDisabled={policy.scope === "privacy-request:read"}
-                  isReadOnly={policy.scope === "privacy-request:read"}
+                  isDisabled={policy.scope === 'privacy-request:read'}
+                  isReadOnly={policy.scope === 'privacy-request:read'}
                 >
                   {policy.privilege}
                 </Checkbox>
@@ -246,19 +246,19 @@ const UserForm: NextPage = () => {
             </Stack>
           </Stack>
 
-          <NextLink href="/user-management" passHref>
-            <Button variant="outline" mr={3} size="sm">
+          <NextLink href='/user-management' passHref>
+            <Button variant='outline' mr={3} size='sm'>
               Cancel
             </Button>
           </NextLink>
           <Button
-            type="submit"
-            bg="primary.800"
-            _hover={{ bg: "primary.400" }}
-            _active={{ bg: "primary.500" }}
-            colorScheme="primary"
+            type='submit'
+            bg='primary.800'
+            _hover={{ bg: 'primary.400' }}
+            _active={{ bg: 'primary.500' }}
+            colorScheme='primary'
             // disabled={!(isValid && dirty)}
-            size="sm"
+            size='sm'
           >
             Save
           </Button>