+++ clients/admin-ui/src/pages/index.tsx
@@ -13,15 +13,15 @@ const Home: NextPage = () => (
   <div>
     <Head>
       <title>Fides Admin UI</title>
-      <meta name="description" content="Generated from FidesUI template" />
-      <link rel="icon" href="/favicon.ico" />
+      <meta name='description' content='Generated from FidesUI template' />
+      <link rel='icon' href='/favicon.ico' />
     </Head>
 
     <NavBar />
 
     <main>
       <Box px={9} py={10}>
-        <Heading mb={8} fontSize="2xl" fontWeight="semibold">
+        <Heading mb={8} fontSize='2xl' fontWeight='semibold'>
           Subject Requests
         </Heading>
         <RequestFilters />